# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "PillMind API",
    "description": "PillMind 3.0 Kurumsal Entegrasyon, HL7 FHIR Uyumlu ve Güvenli İlaç Etkileşim Analiz API Sözleşmesi",
    "version": "3.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Lokal Geliştirme Sunucusu"
    }
  ],
  "paths": {
...
    "/api/pillbox/delete": {
      "delete": {
        "summary": "Kayıtlı İlaç Kutusunu Sil",
        "description": "Kullanıcının belirtilen ID'deki kayıtlı ilaç kutusunu veritabanından siler.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "pillboxId": { "type": "string", "example": "pillbox-uuid-123" }
                },
                "required": ["pillboxId"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Başarılı silme.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true }
                  }
                }
              }
            }
          },
          "401": { "description": "Oturum açılmamış" },
          "404": { "description": "Bulunamadı veya yetkisiz işlem" }
        }
      }
    },
...
  }
}